+++ black_it/utils/seedable.py
@@ -21,8 +21,6 @@ from typing import Optional
 import numpy as np
 from numpy.random import default_rng
 
-from black_it.utils.base import get_random_seed
-
 
 class BaseSeedable:
     """
@@ -69,3 +67,18 @@ class BaseSeedable:
     def get_random_seed(self) -> int:
         """Get new random seed from the current random generator."""
         return get_random_seed(self.__random_generator)
+
+
+def get_random_seed(random_generator: np.random.Generator) -> int:
+    """
+    Get a random seed from a random generator.
+
+    Sample an integer in the range [0, 2^32 - 1].
+
+    Args:
+        random_generator: the random generator to be used for sampling the random seed.
+
+    Returns:
+        the random seed.
+    """
+    return random_generator.integers(2**32 - 1)
