+++ scripts/whitelists/package_whitelist.py
@@ -32,3 +32,4 @@ diff_log_demean_filter  # unused function (black_it/utils/time_series.py:147)
 IDENTITY  # unused variable (black_it/loss_functions/msm.py:43)
 INVERSE_VARIANCE  # unused variable (black_it/loss_functions/msm.py:44)
 LikelihoodLoss  # unused class (black_it/loss_functions/likelihood.py:35)
+BaseSeedable  # unused class (black_it/utils/seedable.py:25)
